+++ packages/ui/resolution/tests/resolution-specs.js
@@ -1,5 +1,6 @@
 import {
 	calculateFontSize,
+	config,
 	defineScreenTypes,
 	getScreenType,
 	init,
@@ -148,6 +149,21 @@ describe('Resolution Specs', () => {
 		expect(actualHD).toBe(expectedHD);
 	});
 
+	test('should convert pixels to units in landscape and fontSizeHandling is "scale"', () => {
+		const closestResolution = {...XGA};
+		const wideResolution = {innerWidth: 1024, innerHeight: 480};
+		init({measurementNode: wideResolution});
+
+		const expectedFHD = 3 / (wideResolution.innerHeight / closestResolution.height) + 'rem';
+		const actualFHD = unit(48, 'rem');
+
+		const expectedHD = 2 / (wideResolution.innerHeight / closestResolution.height) + 'rem';
+		const actualHD = unit(32, 'rem');
+
+		expect(actualFHD).toBe(expectedFHD);
+		expect(actualHD).toBe(expectedHD);
+	});
+
 	test('should select source for the current screen type', function () {
 		let measurementNode;
 		const src = {vga: 'VGA', hd: 'HD', fhd: 'FHD', uhd: 'UHD'};
@@ -210,5 +226,25 @@ describe('Resolution Specs', () => {
 		measurementNode = {innerWidth: UHD2.width, innerHeight: UHD2.height};
 		init({measurementNode});
 		expect(selectSrc(src)).toBe(src.uhd);
+	});
+
+	test('should calculate linearly scaled font size based on workspace bounds and current screen', () => {
+		config.linearScaling.active = true;
+		screenTypes.forEach((type, index, arr) => {
+			init({measurementNode: {innerWidth: type.width, innerHeight: type.height}});
+			let actual = calculateFontSize();
+			expect(actual).toBe(type.pxPerRem + 'px');
+
+			init({measurementNode: {innerWidth: type.width - 10, innerHeight: type.height - 10}});
+			actual = Number(calculateFontSize().split('px').at(0));
+			expect(actual).toBeLessThan(type.pxPerRem);
+
+			init({measurementNode: {innerWidth: type.width + 10, innerHeight: type.height + 10}});
+			actual = Number(calculateFontSize().split('px').at(0));
+
+			const expectCondition = arr.length - 1 === index ? 'toEqual' : 'toBeGreaterThan';
+			expect(actual)[expectCondition](type.pxPerRem);
+		});
+		config.linearScaling.active = false;
 	});
 });